requirements.txt
pylint==2.16.2
mypy==1.4.1
black==23.7.0
pytest==7.2.1
pyjwt==2.4.0
flask==2.2.2
Flask-SQLAlchemy==3.0.5
